DL-12466 update itsetting config and minor changes
