+++ build.sbt
@@ -1,8 +1,11 @@
-import TestPhases.{TemplateItTest, TemplateTest}
 import sbt.Keys.parallelExecution
-import uk.gov.hmrc.DefaultBuildSettings.{addTestReportOption, defaultSettings, scalaSettings}
+import uk.gov.hmrc.DefaultBuildSettings
+import uk.gov.hmrc.DefaultBuildSettings.{defaultSettings, scalaSettings}
 
 val appName: String = "awrs"
+
+ThisBuild / majorVersion := 2
+ThisBuild / scalaVersion := "2.13.12"
 
 lazy val appDependencies : Seq[ModuleID] = AppDependencies()
 lazy val plugins : Seq[Plugins] = Seq(play.sbt.PlayScala, SbtAutoBuildPlugin, SbtGitVersioning, SbtDistributablesPlugin)
@@ -22,26 +25,21 @@ lazy val scoverageSettings = {
 lazy val microservice = Project(appName, file("."))
   .enablePlugins(Seq(play.sbt.PlayScala, SbtDistributablesPlugin) ++ plugins : _*)
   .settings(playSettings ++ scoverageSettings : _*)
-  .settings( majorVersion := 2 )
   .settings(scalaSettings: _*)
   .settings(defaultSettings(): _*)
   .settings(
-    scalaVersion := "2.13.12",
     libraryDependencies ++= appDependencies,
     Test / parallelExecution := false,
     retrieveManaged := true,
-    scalacOptions ++= Seq("-feature"),
+    scalacOptions ++= Seq("-feature", "-Wconf:src=routes/.*:s"),
   )
-  .settings(inConfig(TemplateTest)(Defaults.testSettings): _*)
-  .settings(inConfig(TemplateItTest)(Defaults.itSettings): _*)
-  .configs(IntegrationTest)
-  .settings(
-    IntegrationTest / Keys.fork := true,
-    IntegrationTest / unmanagedSourceDirectories := (IntegrationTest / baseDirectory)(base => Seq(base / "it")).value,
-    addTestReportOption(IntegrationTest, "int-test-reports"),
-    IntegrationTest / parallelExecution:= false,
-    scalacOptions += "-Wconf:src=routes/.*:s")
   .settings(
     resolvers += Resolver.jcenterRepo
   )
   .disablePlugins(JUnitXmlReportPlugin)
+
+lazy val it = project
+  .enablePlugins(PlayScala)
+  .dependsOn(microservice % "test->test") // the "test->test" allows reusing test code and test dependencies
+  .settings(DefaultBuildSettings.itSettings())
+  .settings(libraryDependencies ++= AppDependencies.itDependencies)
