+++ it/test/uk/gov/hmrc/helpers/IntegrationSpec.scala
@@ -1,7 +1,22 @@
+/*
+ * Copyright 2024 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
 package uk.gov.hmrc.helpers
 
-import akka.util.Timeout
+import org.apache.pekko.util.Timeout
 import org.scalatest._
 import org.scalatest.wordspec.AnyWordSpecLike
 import play.api.libs.ws.WSRequest
@@ -10,7 +25,6 @@ import uk.gov.hmrc.helpers.application.IntegrationApplication
 import uk.gov.hmrc.helpers.http.StubbedBasicHttpCalls
 import uk.gov.hmrc.helpers.wiremock.WireMockSetup
 import uk.gov.hmrc.http.HeaderCarrier
-
 import scala.concurrent.Future
 import scala.concurrent.duration._
