--- it/uk/gov/hmrc/helpers/wiremock/WireMockConfig.scala
@@ -1,8 +1,0 @@
-
-package uk.gov.hmrc.helpers.wiremock
-
-trait WireMockConfig {
-  val wireMockPort: Int    = 11111
-  val wireMockHost: String = "localhost"
-  val wireMockUrl: String  = s"http://$wireMockPort:$wireMockPort"
-}
